@@ -16,7 +16,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.1-pre+1
+version: 2.0.1-pre+1
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
@@ -37,44 +37,43 @@ dependencies:
   dio: ^5.4.1
   cookie_jar: ^4.0.8
   dio_cookie_manager: ^3.1.1
-  # 图片
-  path_provider: ^2.1.1
+  path_provider: ^2.1.5
 
   # 设备信息
-  device_info_plus: ^9.0.0
-  font_awesome_flutter: ^10.4.0
+  device_info_plus: ^12.3.0
+  font_awesome_flutter: ^10.12.0
   # toast
   flutter_smart_dialog: ^4.9.4
   # 加密
-  crypto: ^3.0.3
+  crypto: ^3.0.7
   encrypt: ^5.0.3
 
   # 进度条
   audio_video_progress_bar: ^2.0.3
-  auto_orientation: ^2.3.1
+  auto_orientation_v2: ^2.3.7
 
   # 获取appx信息
-  package_info_plus: ^4.2.0
   url_launcher: ^6.2.6
 
-  uuid: ^3.0.7
+  uuid: ^4.5.2
   logger: ^2.3.0
   path: ^1.9.0
   # 二维码
   qr_flutter: ^4.1.0
   loading_more_list: ^7.1.0
   gt3_flutter_plugin: ^0.1.0
-  go_router: ^14.8.1
-  flutter_native_splash: ^2.4.4
+  go_router: ^17.0.1
+  flutter_native_splash: ^2.4.7
   provider: ^6.1.4
-  intl: ^0.19.0
+  intl: ^0.20.2
   pager: ^0.0.6
   sqflite: ^2.3.3
   shared_preferences: ^2.5.3
   window_manager: ^0.5.1
-  media_kit_video: ^2.0.1
   media_kit: ^1.2.6
+  media_kit_video: ^2.0.1
   media_kit_libs_video: ^1.0.7
+  custom_radio_grouped_button: ^2.3.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
