@@ -1,7 +1,21 @@
+import 'dart:io';
 import 'package:flutter/material.dart';
+import 'package:custom_radio_grouped_button/custom_radio_grouped_button.dart';
+import 'package:provider/provider.dart';
 
 import '/widgets/wallclock_wrapper.dart';
 import './tabconf.dart';
+
+class GradeNotifier extends ChangeNotifier {
+  String _selectedGrade = 'primary_upper';
+
+  String get selectedGrade => _selectedGrade;
+
+  void setGrade(String grade) {
+    _selectedGrade = grade;
+    notifyListeners();
+  }
+}
 
 class StudyVideoPage extends StatefulWidget {
   const StudyVideoPage({super.key});
@@ -15,6 +29,7 @@ class _StudyVideoState extends State<StudyVideoPage>
   late final TabController _tabController;
   late final List<Widget> _tabPages;
   int _currentIndex = 0;
+  final GradeNotifier _gradeNotifier = GradeNotifier();
 
   @override
   void initState() {
@@ -30,17 +45,22 @@ class _StudyVideoState extends State<StudyVideoPage>
   @override
   void dispose() {
     _tabController.dispose();
+    _gradeNotifier.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return WallClockWrapper(
-      child: Scaffold(
-        extendBody: true,
-        extendBodyBehindAppBar: true,
-        backgroundColor: Colors.transparent,
-        body: _buildContent(),
+    return ChangeNotifierProvider.value(
+      value: _gradeNotifier,
+      child: WallClockWrapper(
+        child: Scaffold(
+          extendBody: true,
+          extendBodyBehindAppBar: true,
+          backgroundColor: Colors.transparent,
+          body: _buildContent(),
+          floatingActionButton: _buildFloatingActionButton(),
+        ),
       ),
     );
   }
@@ -82,19 +102,55 @@ class _StudyVideoState extends State<StudyVideoPage>
           enableFeedback: true,
           splashBorderRadius: BorderRadius.circular(10),
           tabAlignment: TabAlignment.center,
-          onTap: _onTabTapped,
+          onTap: (index) {
+            _currentIndex = index;
+          },
         ),
       ),
     );
   }
 
-  void _onTabTapped(int index) {
-    setState(() {
-      if (_currentIndex == index) {
-        // 如果点击的是当前选中的标签，可以执行刷新操作
-        // 这里可以根据需要添加具体逻辑
-      }
-      _currentIndex = index;
-    });
+  Widget _buildFloatingActionButton() {
+    return Container(
+      width: 62,
+      margin: Platform.isAndroid || Platform.isIOS
+          ? const EdgeInsets.only(bottom: 66.0)
+          : const EdgeInsets.only(bottom: 16.0),
+      child: CustomRadioButton<String>(
+        elevation: 0,
+        horizontal: true,
+        absoluteZeroSpacing: false,
+        unSelectedColor: Colors.transparent,
+        selectedColor: Colors.blue,
+        buttonLables: const ['小初', '小高', '初中'],
+        buttonValues: const ['primary', 'primary_upper', 'junior'],
+        defaultSelected: _gradeNotifier.selectedGrade,
+        radioButtonValue: (value) {
+          if (value != null) {
+            _gradeNotifier.setGrade(value);
+          }
+        },
+        width: 58,
+        height: 28,
+        enableShape: true,
+        shapeRadius: 8,
+        margin: const EdgeInsets.all(0),
+        selectedBorderColor: Colors.blue,
+        unSelectedBorderColor: Colors.grey.withValues(alpha: 0.5),
+        buttonTextStyle: const ButtonTextStyle(
+          selectedColor: Colors.white,
+          unSelectedColor: Colors.grey,
+          textStyle: TextStyle(
+            fontSize: 11,
+            fontWeight: FontWeight.w500,
+          ),
+          selectedTextStyle: TextStyle(
+            fontSize: 11,
+            color: Colors.white,
+            fontWeight: FontWeight.w600,
+          ),
+        ),
+      ),
+    );
   }
 }
