@@ -2,63 +2,39 @@ import 'tabview.dart';
 
 List tabsConfig = [
   {
-    'label': '全部',
-    'page': const StudyTabView(
-      subjectName: 'all',
-    ),
-  },
-  {
     'label': '英语',
-    'page': const StudyTabView(
-      subjectName: 'english',
-    ),
+    'page': const StudyTabView(subjectName: 'english'),
   },
   {
     'label': '语文',
-    'page': const StudyTabView(
-      subjectName: 'chinese',
-    ),
+    'page': const StudyTabView(subjectName: 'chinese'),
   },
   {
     'label': '数学',
-    'page': const StudyTabView(
-      subjectName: 'math',
-    ),
+    'page': const StudyTabView(subjectName: 'math'),
   },
   {
     'label': '物理',
-    'page': const StudyTabView(
-      subjectName: 'physics',
-    ),
+    'page': const StudyTabView(subjectName: 'physics'),
   },
   {
     'label': '化学',
-    'page': const StudyTabView(
-      subjectName: 'chemistry',
-    ),
+    'page': const StudyTabView(subjectName: 'chemistry'),
   },
   {
     'label': '生物',
-    'page': const StudyTabView(
-      subjectName: 'biology',
-    ),
+    'page': const StudyTabView(subjectName: 'biology'),
   },
   {
     'label': '地理',
-    'page': const StudyTabView(
-      subjectName: 'geography',
-    ),
+    'page': const StudyTabView(subjectName: 'geography'),
   },
   {
     'label': '历史',
-    'page': const StudyTabView(
-      subjectName: 'history',
-    ),
+    'page': const StudyTabView(subjectName: 'history'),
   },
   {
     'label': '政治',
-    'page': const StudyTabView(
-      subjectName: 'politics',
-    ),
+    'page': const StudyTabView(subjectName: 'politics'),
   },
 ];
